+++ config/gitlab.php
@@ -9,6 +9,8 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 return [
 
     /*
@@ -39,12 +41,12 @@ return [
 
         'main' => [
             'token' => 'your-token',
-            'base_url' => 'http://git.yourdomain.com/api/v3/',
+            'url' => 'https://git.yourdomain.com',
         ],
 
         'alternative' => [
             'token' => 'your-token',
-            'base_url' => 'http://git.yourdomain.com/api/v3/',
+            'url' => 'https://git.yourdomain.com',
         ],
 
     ],