+++ src/GitLabFactory.php
@@ -9,26 +9,28 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace Vinkla\GitLab;
 
 use Gitlab\Client;
 use InvalidArgumentException;
 
 /**
- * This is the GitLab factory class.
+ * This is the gitlab factory class.
  *
  * @author Vincent Klaiber <hello@vinkla.com>
  */
 class GitLabFactory
 {
     /**
-     * Make a new GitLab client.
+     * Make a new gitlab client.
      *
      * @param array $config
      *
      * @return \Gitlab\Client
      */
-    public function make(array $config)
+    public function make(array $config): Client
     {
         $config = $this->getConfig($config);
 
@@ -42,11 +44,11 @@ class GitLabFactory
      *
      * @throws \InvalidArgumentException
      *
-     * @return string
+     * @return array
      */
-    protected function getConfig(array $config)
+    protected function getConfig(array $config): array
     {
-        $keys = ['token', 'base_url'];
+        $keys = ['token', 'url'];
 
         foreach ($keys as $key) {
             if (!array_key_exists($key, $config)) {
@@ -54,7 +56,7 @@ class GitLabFactory
             }
         }
 
-        return array_only($config, ['token', 'base_url', 'method', 'sudo']);
+        return array_only($config, ['token', 'url', 'method', 'sudo']);
     }
 
     /**
@@ -64,13 +66,13 @@ class GitLabFactory
      *
      * @return \Gitlab\Client
      */
-    protected function getClient(array $config)
+    protected function getClient(array $config): Client
     {
-        $client = new Client($config['base_url']);
+        $client = Client::create($config['url']);
 
         $client->authenticate(
             $config['token'],
-            array_get($config, 'method', Client::AUTH_HTTP_TOKEN),
+            array_get($config, 'method', Client::AUTH_URL_TOKEN),
             array_get($config, 'sudo', null)
         );