+++ src/GitLabServiceProvider.php
@@ -9,6 +9,8 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace Vinkla\GitLab;
 
 use Gitlab\Client;
@@ -116,7 +118,7 @@ class GitLabServiceProvider extends ServiceProvider
      *
      * @return string[]
      */
-    public function provides()
+    public function provides(): array
     {
         return [
             'gitlab',