+++ tests/GitLabFactoryTest.php
@@ -9,6 +9,8 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace Vinkla\Tests\GitLab;
 
 use Gitlab\Client;
@@ -27,7 +29,7 @@ class GitLabFactoryTest extends AbstractTestCase
 
         $return = $factory->make([
             'token' => 'your-token',
-            'base_url' => 'http://git.yourdomain.com/api/v3/',
+            'url' => 'https://git.yourdomain.com',
         ]);
 
         $this->assertInstanceOf(Client::class, $return);
@@ -41,14 +43,14 @@ class GitLabFactoryTest extends AbstractTestCase
         $factory = $this->getGitLabFactory();
 
         $factory->make([
-            'base_url' => 'http://git.yourdomain.com/api/v3/',
+            'url' => 'https://git.yourdomain.com',
         ]);
     }
 
     /**
      * @expectedException \InvalidArgumentException
      */
-    public function testMakeWithoutBaseUrl()
+    public function testMakeWithoutUrl()
     {
         $factory = $this->getGitLabFactory();