+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # CHANGELOG
 
+## 4.1.0 (released 2017-09-02)
+
+- Removed laravel 5.3 and 5.4 support
+
+## 4.0.0 (released 2017-08-23)
+
+- Added gitlab api 4.0 support
+- Added httplug support
+- Updated base_url config to url
+
+## 3.0.0 (released 2017-08-07)
+
+- Updated gitlab package
+- Added laravel 5.5 support
+
+## 2.6.0 (released 2017-01-01)
+
+- Added laravel 5.4 support
+- Dropped php 5.6 support
+
 ## 2.5.0 (released 2016-07-11)
 
 - Dropped php 5.5 support